+++ cylc/flow/scripts/jobs_kill.py
@@ -21,8 +21,8 @@
 Read job status files to obtain the names of the job runners and the job IDs
 in the runners. Invoke the relevant job runner commands to ask the job runners
 to terminate the jobs.
-
 """
+
 from cylc.flow.job_runner_mgr import JobRunnerManager
 from cylc.flow.option_parsers import CylcOptionParser as COP
 from cylc.flow.terminal import cli_function