+++ tests/functional/spawn-on-demand/16-parents-yes-no.t
@@ -1,6 +1,7 @@
+#!/usr/bin/env bash
 # THIS FILE IS PART OF THE CYLC WORKFLOW ENGINE.
 # Copyright (C) NIWA & British Crown (Met Office) & Contributors.
-#
+# 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
@@ -13,12 +14,13 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""Tests for Cylc scheduler CLI."""
+#-------------------------------------------------------------------------------
 
-from cylc.flow.scheduler_cli import _protect_remote_cmd_args
+# A task with no parents should auto-spawn out to the runahead limit, except
+# in cycle points where it does have parents. And those cycles, if any, should
+# not block spawning of subsequent parentless cycles - GitHub #4906.
 
-
-def test__protect_remote_cmd_args():
-    cmd = ['cylc', 'play', '-n', '--set=FOO="foo"', 'wf']
-    exp = ['cylc', 'play', '-n', '\'--set=FOO="foo"\'', 'wf']
-    assert _protect_remote_cmd_args(cmd) == exp
+. "$(dirname "$0")/test_header"
+set_test_number 2
+reftest
+exit